refactor: 테스트 파일 위치 수정
