+++ src/hooks/useCardForm/index.ts
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { filterDigits } from "../utils";
+import { filterDigits } from "../../utils";
 
 const useCardForm = () => {
   const [cardNumber, setCardNumber] = useState(["", "", "", ""]);
